Establish new default rate limit
This commit establishes a new default rate limit for the
summary_by_domain task. Since the rate limit is applied per worker, I
reduced the limit to 2 tasks per hour and the number of cert-worker
replicas to 5 (for easy math). This way, we can match the rate limit for
the Small Tier of the Certificate Transparency Search API.